holo/packages/img2mesh3d/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img2mesh3d"
version = "0.2.0"
description = "2D image to 3D pipeline toolkit (Replicate multi-view + depth + Meshy reconstruction) with AWS SQS job runner"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "img2mesh3d contributors" }]
keywords = ["3d", "meshy", "replicate", "threejs", "aws", "sqs", "dynamodb", "s3"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "boto3>=1.28.0",
  "pillow>=10.0.0",
  "pydantic>=2.0.0",
  "requests>=2.31.0",
  "rich>=13.0.0",
  "tenacity>=8.2.0",
  "typer>=0.12.0",
  "replicate>=1.0.0",
]

[project.optional-dependencies]
api = [
  "fastapi>=0.110.0",
  "uvicorn>=0.23.0",
  "python-multipart>=0.0.9",
]
dev = [
  "pytest>=8.0.0",
  "moto[s3,sqs,dynamodb]>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.8.0",
  "types-requests>=2.31.0",
]

[project.scripts]
img2mesh3d = "img2mesh3d.cli:app"
img2mesh3d-worker = "img2mesh3d.worker_main:main"

[tool.setuptools]
package-dir = { "img2mesh3d" = "src" }
packages = [
  "img2mesh3d",
  "img2mesh3d.api",
  "img2mesh3d.aws",
  "img2mesh3d.clients",
  "img2mesh3d.jobs",
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
